populating users and posts and deleteing their relations when post is deleted

+++ server/controllers/feed.js
@@ -2,6 +2,7 @@ const { validationResult } = require("express-validator");
 const path = require("path");
 const { deleteFile } = require("../utils/clearImage");
 const Post = require("../models/PostModel");
+const User = require("../models/UserModel");
 exports.getPosts = async (req, res, next) => {
   try {
     const currentPage = req.query.page || 1;
@@ -10,7 +11,8 @@ exports.getPosts = async (req, res, next) => {
     totalItems = await Post.find().countDocuments();
     const posts = await Post.find()
       .skip(perPage * (currentPage - 1))
-      .limit(perPage);
+      .limit(perPage)
+      .populate({ path: "creator" });
     if (posts.length === 0) {
       const error = new Error("Could not fetch posts");
       error.statusCode = 404;
@@ -32,7 +34,8 @@ exports.getPosts = async (req, res, next) => {
 exports.getPost = async (req, res, next) => {
   try {
     const { postId } = req.params;
-    const post = await Post.findById(postId);
+    const post = await Post.findById(postId).populate({ path: "creator" });
+    const user = await User.findById(req.userId);
     if (!post) {
       const error = new Error("Could not find post");
       error.statusCode = 404;
@@ -41,6 +44,7 @@ exports.getPost = async (req, res, next) => {
     res.status(200).json({
       message: "Post fetched",
       post: post,
+      creator: user.name,
     });
   } catch (err) {
     if (!err.statusCode) {
@@ -66,19 +70,25 @@ exports.createPost = async (req, res, next) => {
     const title = req.body.title;
     const content = req.body.content;
     const imageUrl = req.file.filename;
+    const user = await User.findById(req.userId);
     const post = await Post.create({
       title: title,
       content: content,
       creator: {
-        name: "Max",
+        _id: req.userId,
       },
       imageUrl: imageUrl,
     });
 
-    // Create post in db
+    user.posts.push(post); // Create post in db
+    await user.save();
     res.status(201).json({
       message: "Post created successfully!",
       post: post,
+      creator: {
+        _id: user._id,
+        name: user.name,
+      },
     });
   } catch (err) {
     if (!err.statusCode) {
@@ -108,10 +118,16 @@ exports.updatePost = async (req, res, next) => {
       error.statusCode = 422;
       return next(error);
     }
+    console.log(postId);
     const post = await Post.findById(postId);
     if (!post) {
       const error = new Error("Could not find post");
       error.statusCode = 422;
+      return next(error);
+    }
+    if (post.creator._id.toString() !== req.userId) {
+      const error = new Error("You dont have permission");
+      error.statusCode = 403;
       return next(error);
     }
     if (imageUrl !== post.imageUrl) {
@@ -139,6 +155,11 @@ exports.deletePost = async (req, res, next) => {
     const { postId } = req.params;
     const post = await Post.findById(postId);
     // check if creator is logged in user
+    if (post.creator._id.toString() !== req.userId) {
+      const error = new Error("You dont have permission");
+      error.statusCode = 403;
+      return next(error);
+    }
     deleteFile(post.imageUrl);
     if (!post) {
       const error = new Error("Post does not exists");
@@ -146,6 +167,9 @@ exports.deletePost = async (req, res, next) => {
       next(error);
     }
     const removedPost = await Post.findByIdAndRemove(postId);
+    const user = await User.findById(req.userId);
+    user.posts.pull(postId);
+    await user.save();
     res.status(200).json({
       message: "Post was deleted",
     });
